basic UI Toolkit ProgressBar

+++ Runtime/ProgressBarAttribute.cs
@@ -7,11 +7,11 @@ namespace SaintsField
         public SaintsAttributeType AttributeType => SaintsAttributeType.Field;
         public string GroupBy => "__LABEL_FIELD__";
 
-        public readonly double Min;
+        public readonly float Min;
         public readonly string MinCallback;
-        public readonly double Max;
+        public readonly float Max;
         public readonly string MaxCallback;
-        public readonly double Step;
+        public readonly float Step;
 
         public readonly EColor Color;
         public readonly string ColorCallback;
@@ -22,25 +22,25 @@ namespace SaintsField
         public readonly string TitleCallback;
 
         public ProgressBarAttribute(): this(0, 100, -1) { }
-        public ProgressBarAttribute(double max): this(0, max, -1) { }
-        public ProgressBarAttribute(double max, double step): this(0, max, step) { }
+        public ProgressBarAttribute(float max): this(0, max, -1) { }
+        public ProgressBarAttribute(float max, float step): this(0, max, step) { }
 
         public ProgressBarAttribute(string maxCallback): this(0, maxCallback) { }
-        public ProgressBarAttribute(string maxCallback, double step): this(0, maxCallback, step) { }
+        public ProgressBarAttribute(string maxCallback, float step): this(0, maxCallback, step) { }
 
-        public ProgressBarAttribute(string maxCallback = null, double step = -1, EColor color = EColor.OceanicSlate,
+        public ProgressBarAttribute(string maxCallback = null, float step = -1, EColor color = EColor.OceanicSlate,
             EColor backgroundColor = EColor.CharcoalGray, string colorCallback = null,
             string backgroundColorCallback = null)
             : this(0, maxCallback, step, color, backgroundColor, colorCallback, backgroundColorCallback)
         {
         }
 
-        public ProgressBarAttribute(double max=100, double step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null)
+        public ProgressBarAttribute(float max=100, float step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null)
             : this(0, max, step, color, backgroundColor, colorCallback, backgroundColorCallback)
         {
         }
 
-        public ProgressBarAttribute(double min=0, double max=100, double step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null, string titleCallback=null)
+        public ProgressBarAttribute(float min=0, float max=100, float step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null, string titleCallback=null)
         {
             Min = min;
             MinCallback = null;
@@ -54,7 +54,7 @@ namespace SaintsField
             TitleCallback = titleCallback;
         }
 
-        public ProgressBarAttribute(double min=0, string maxCallback=null, double step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null, string titleCallback=null)
+        public ProgressBarAttribute(float min=0, string maxCallback=null, float step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null, string titleCallback=null)
         {
             Min = min;
             MinCallback = null;
@@ -68,7 +68,7 @@ namespace SaintsField
             TitleCallback = titleCallback;
         }
 
-        public ProgressBarAttribute(string minCallback=null, double max=100, double step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null, string titleCallback=null)
+        public ProgressBarAttribute(string minCallback=null, float max=100, float step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null, string titleCallback=null)
         {
             // Min = min;
             MinCallback = minCallback;
@@ -82,7 +82,7 @@ namespace SaintsField
             TitleCallback = titleCallback;
         }
 
-        public ProgressBarAttribute(string minCallback=null, string maxCallback=null, double step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null, string titleCallback=null)
+        public ProgressBarAttribute(string minCallback=null, string maxCallback=null, float step=-1, EColor color=EColor.OceanicSlate, EColor backgroundColor=EColor.CharcoalGray, string colorCallback=null, string backgroundColorCallback=null, string titleCallback=null)
         {
             // Min = min;
             MinCallback = minCallback;
